requirements.txt
Flask==3.0.3
gunicorn==22.0.0
line-bot-sdk==3.19.0
requests==2.32.3
beautifulsoup4==4.12.3